Fix Time addition, also make Time's min double

+++ 5black/5week/sprav_t/working_time.h
@@ -11,5 +11,5 @@ struct Time {
   Time operator+(double additional_min) const;
 
   uint8_t day = 0;
-  uint16_t min = 0; 
+  double min = 0; 
 };
